Some changes to the manifest, adding some activities

+++ Team2RecipeFinder/src/ca/ualberta/team2recipefinder/Controller.java
@@ -27,7 +27,7 @@ public class Controller {
 		return model.searchRecipe(keywords);
 	}
 	
-	public Recipe getRecipe(int id) {
+	public Recipe getRecipe(long id) {
 		return model.getRecipeById(id);
 	}
